+++ src/app/components/Card/__tests__/mocks.js
@@ -1,0 +1,10 @@
+const defaultMock = {
+    brand: 'Hp',
+	imgUrl: 'http:image-hp-test.com',
+	model: 'test model',
+	price: '110',
+};
+
+module.exports = {
+    defaultMock
+};