+++ CHANGELOG.md
@@ -9,3 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create scaffolding for the new techtronix-store app.
 - Added initial configurations to unit tests, linter and husky.
 - Project pipeline is created to deploy the application with githubActions.
+
+## [Unreleased]
+### Added
+- Products view.
+- Product list service. PR [#1](https://github.com/jpaomat/ttx-store-frontend/pull/1)